requirements.txt
Flask
docx
Gunicorn
sqlalchemy
pymysql
xhtml2pdf
babel
docxtpl
flask_login
Flask-WTF
pandas
openpyxl
plotly
kaleido==0.2.1
PyPDF2
google-generativeai
google-api-python-client
google-auth
google-auth-httplib2
pdfplumber
gspread
pypdf
PyMuPDF
google-genai
openai
pycairo==1.27.0
Flask
docx
Gunicorn
sqlalchemy
pymysql
xhtml2pdf
babel
docxtpl
flask_login
pandas
openpyxl
plotly
kaleido==0.2.1
PyPDF2
google-generativeai
google-api-python-client
google-auth
google-auth-httplib2
pdfplumber
gspread
pypdf
PyMuPDF
google-genai
openai
pycairo==1.27.0
python-dotenv